+++ internal/web/web_test.go
@@ -13,7 +13,6 @@ import (
 
 	"github.com/ziyan/cue/internal/browser"
 	"github.com/ziyan/cue/internal/config"
-	"github.com/ziyan/cue/internal/fleet"
 	"github.com/ziyan/cue/internal/network"
 	"github.com/ziyan/cue/internal/supervise"
 	"github.com/ziyan/cue/internal/timesync"
@@ -57,7 +56,6 @@ func (self *fakeDevice) VNCAddress() string                    { return "127.0.0
 func (self *fakeDevice) StartedAt() time.Time                  { return self.startedAt }
 func (self *fakeDevice) XServer() *xserver.Server              { return self.xserver }
 func (self *fakeDevice) TimeSync() *timesync.Client            { return self.timesync }
-func (self *fakeDevice) Fleet() *fleet.Tunnel                  { return nil }
 func (self *fakeDevice) Network() *network.Manager             { return nil }
 func (self *fakeDevice) Restart(context.Context, string) error { return nil }